[test] Move RecordTransformer tests from DaVinciClientTest to DaVinciClientRecordTransformerTest
