+++ internal/venice-test-common/build.gradle
@@ -178,7 +178,8 @@ def integrationTestBuckets = [
       "com.linkedin.venice.endToEnd.TestVson*",
       "com.linkedin.venice.endToEnd.Push*"],
   "1210": [
-      "com.linkedin.venice.hadoop.*"],
+      "com.linkedin.venice.hadoop.*",
+      "com.linkedin.venice.endToEnd.DaVinciClientRecordTransformerTest"],
   "1220": [
       "com.linkedin.venice.endToEnd.TestPushJob*"],
   "1230": [
